Finished a version of the abstract base NS class and got all the other Nested Samplers working as derived classes from the base.
